refactoring shell.OutputWriter faker
